Add support for purging RaidList, winnerlist, and oldlist as well.
This will purge based on the last X raids used for attendance
calculation.  Anything older will be either concatenated or disposed
of.  This will take care of limiting the amount of history retained for
quickly growing data.  Be sure to back up your fRaid.lua prior to using
this.

For winnerlist and oldlist to be purged correctly, the timestamps need
to be corrected.  An updated fList/fLib will write new timestamps
correctly (in year-first format), but you should call LIST.FixTime() to
fix existing data before the first purge.  Currently, it just discards
all winnerlist and oldlist data, but could be changed (if anyone cares)
to fix the date strings.

git-svn-id: https://tamaddons.googlecode.com/svn/trunk/fRaid@246 7aec42f0-9e3d-11dd-ada7-d1ebab521e76