Vote for a player in the raid support.
Allows the raid to vote for a player in the raid.  Useful for
distributing rare but useless loot like Invincible.

New commands:

votestart: Starts a vote, announces to the raid

vote (player): Votes for the given player, the voter and the votee must
both be in the raid.

abstain: Abstain from voting, but acknowledge that you chose to abstain
and aren't just AFK.

voteend: Ends the vote, announces the winner to the raid and the voting
results to the officer channel, and copies the vote data to the
'oldvote' section of the LUA data.

git-svn-id: https://tamaddons.googlecode.com/svn/trunk/fRaid@250 7aec42f0-9e3d-11dd-ada7-d1ebab521e76